添加一些快捷发信息的函数

+++ model.go
@@ -290,7 +290,7 @@ type GroupAdminSysNotifyPack struct {
 		Flag8           int    `json:"Flag_8"`
 		GroupID         int64  `json:"GroupId"`
 		GroupName       string `json:"GroupName"`
-		ActionUin       int    `json:"ActionUin"`
+		ActionUin       int64  `json:"ActionUin"`
 		ActionName      string `json:"ActionName"`
 		ActionGroupCard string `json:"ActionGroupCard"`
 		Action          int    `json:"Action"`
@@ -373,7 +373,7 @@ type UserCardInfo struct {
 type FriendFileResult struct {
 	FileName string `json:"FileName"`
 	FileSize int    `json:"FileSize"`
-	FromUin  int    `json:"FromUin"`
+	FromUin  int64  `json:"FromUin"`
 	URL      string `json:"Url"`
 }
 type GroupFileResult struct {
@@ -381,9 +381,9 @@ type GroupFileResult struct {
 	URL string `json:"Url"`
 }
 type GroupMemberList struct {
-	Count      int `json:"Count"`
-	GroupUin   int `json:"GroupUin"`
-	LastUin    int `json:"LastUin"`
+	Count      int   `json:"Count"`
+	GroupUin   int64 `json:"GroupUin"`
+	LastUin    int64 `json:"LastUin"`
 	MemberList []struct {
 		Age           int    `json:"Age"`
 		AutoRemark    string `json:"AutoRemark"`
@@ -396,7 +396,7 @@ type GroupMemberList struct {
 		JoinTime      int    `json:"JoinTime"`
 		LastSpeakTime int    `json:"LastSpeakTime"`
 		MemberLevel   int    `json:"MemberLevel"`
-		MemberUin     int    `json:"MemberUin"`
+		MemberUin     int64  `json:"MemberUin"`
 		Memo          string `json:"Memo"`
 		NickName      string `json:"NickName"`
 		ShowName      string `json:"ShowName"`
@@ -409,7 +409,7 @@ type GroupMemberList struct {
 type FriendList struct {
 	FriendCount int `json:"Friend_count"`
 	Friendlist  []struct {
-		FriendUin int    `json:"FriendUin"`
+		FriendUin int64  `json:"FriendUin"`
 		IsRemark  bool   `json:"IsRemark"`
 		NickName  string `json:"NickName"`
 		OnlineStr string `json:"OnlineStr"`
@@ -426,8 +426,8 @@ type GroupList struct {
 	Count     int    `json:"Count"`
 	NextToken string `json:"NextToken"`
 	TroopList []struct {
-		GroupID          int    `json:"GroupId"`
-		GroupMemberCount int    `json:"GroupMemberCount"`
+		GroupID          int64  `json:"GroupId"`
+		GroupMemberCount int64  `json:"GroupMemberCount"`
 		GroupName        string `json:"GroupName"`
 		GroupNotice      string `json:"GroupNotice"`
 		GroupOwner       int64  `json:"GroupOwner"`
@@ -457,7 +457,7 @@ type UserInfo struct {
 		Qzone         int    `json:"qzone"`
 		Realname      string `json:"realname"`
 		Smartname     string `json:"smartname"`
-		Uin           int    `json:"uin"`
+		Uin           int64  `json:"uin"`
 	} `json:"data"`
 	Default int    `json:"default"`
 	Message string `json:"message"`
